Issue-3499: Move progress_stream module to compose/cli package
Partially implements #3499
