+++ compose/cli/progress_stream.py
