+++ tests/integration/testcases.py
@@ -8,6 +8,7 @@ from pytest import skip
 
 from .. import unittest
 from compose.cli.docker_client import docker_client
+from compose.cli.progress_stream import stream_output
 from compose.common.const import API_VERSIONS
 from compose.common.const import LABEL_PROJECT
 from compose.config.config import resolve_environment
@@ -15,7 +16,6 @@ from compose.config.config import V1
 from compose.config.config import V2_0
 from compose.config.environment import Environment
 from compose.core import dockerclient as dc
-from compose.progress_stream import stream_output
 from compose.service import Service
